+++ BobsPanicBox/Module_BobsPanicBox.cs
@@ -17,7 +17,7 @@ namespace BobsPanicBox
 
  
         [KSPField(isPersistant = true)]
-        internal float disableAfter = 600;
+        internal int disableAfter = 600;
 
         [KSPField(isPersistant = true)]
         internal int actionAfterTimeout = 0; 
@@ -59,7 +59,7 @@ namespace BobsPanicBox
         internal int postAbortAction;
 
         [KSPField(isPersistant = true)]
-        internal float postAbortDelay;
+        internal int postAbortDelay;
 
         [KSPField(isPersistant = true)]
         internal bool delayPostAbortUntilSafe;
